chore: split cdn test into separate suite
