+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "tsc",
     "lint": "eslint src test --max-warnings 0 && prettier src test json --check",
-    "test": "ts-mocha -u tdd test/**/*.spec.ts --recursive --timeout 10000",
+    "test": "ts-mocha -u tdd 'test/**/*.spec.ts' -exclude test/cdn.spec.ts --recursive --timeout 10000",
+    "test:cdn": "ts-mocha -u tdd 'test/cdn.spec.ts' --timeout 10000",
     "count": "node prod/counts.js",
     "list-different": "node prod/listDifferent.js",
     "audit-outfits": "node prod/auditOutfits.js",
