+++ test/data.spec.ts
@@ -1,7 +1,5 @@
 import { assert } from "chai";
-import { before } from "mocha";
 import { findBestMatch } from "string-similarity";
-import { Parser } from "xml2js";
 
 import {
   AdventureData,
@@ -10,10 +8,6 @@ import {
   PortraitsData,
   PosesData,
 } from "../src/interfaces/Data";
-import { FileList } from "../src/interfaces/FileList";
-
-// eslint-disable-next-line init-declarations
-let fileList: FileList[];
 
 const adventureList = ["becca", "naomi", "rosalia"];
 const emoteList = ["becca", "naomi"];
@@ -21,37 +15,8 @@ const outfitList = ["naomi"];
 const portraitsList = ["becca", "beccalia", "naomi", "rosalia"];
 const posesList = ["becca", "beccalia", "naomi", "novas", "rosalia", "melody"];
 
-const getFiles = async () => {
-  const raw = await fetch(`https://cdn.naomi.lgbt`);
-  const text = await raw.text();
-  const parser = new Parser();
-  const parsed = await parser.parseStringPromise(text);
-  return parsed.ListBucketResult.Contents as FileList[];
-};
-
 suite("Asset data", () => {
-  before(async () => {
-    fileList = await getFiles();
-  });
-
   for (const ns of adventureList) {
-    test(`should have all adventures for ${ns}`, async () => {
-      const json = (await import(`../json/${ns}/adventures.json`))
-        .default as AdventureData[];
-      const data = fileList
-        .filter((f) => f.Key[0].startsWith(`${ns}/games`))
-        .map((i) => i.Key[0].split("/")[2])
-        .filter((el) => el);
-      for (const file of data) {
-        const adventure = json.find((a) => a.fileName === file);
-        assert.isDefined(adventure, `JSON is missing ${file}`);
-      }
-      for (const object of json) {
-        assert.include(data, object.fileName);
-      }
-      assert.equal(data.length, json.length);
-    });
-
     test(`should have unique names and filenames for ${ns}`, async () => {
       const json = (await import(`../json/${ns}/adventures.json`))
         .default as AdventureData[];
@@ -93,23 +58,6 @@ suite("Asset data", () => {
   }
 
   for (const ns of emoteList) {
-    test(`should have all emotes for ${ns}`, async () => {
-      const json = (await import(`../json/${ns}/emotes.json`))
-        .default as EmoteData[];
-      const data = fileList
-        .filter((f) => f.Key[0].startsWith(`${ns}/emotes`))
-        .map((i) => i.Key[0].split("/")[2])
-        .filter((el) => el);
-      for (const file of data) {
-        const emote = json.find((a) => a.fileName === file);
-        assert.isDefined(emote, `JSON is missing ${file}`);
-      }
-      for (const object of json) {
-        assert.include(data, object.fileName);
-      }
-      assert.equal(data.length, json.length);
-    });
-
     test(`should have unique names and filenames for ${ns}`, async () => {
       const json = (await import(`../json/${ns}/emotes.json`))
         .default as EmoteData[];
@@ -151,23 +99,6 @@ suite("Asset data", () => {
   }
 
   for (const ns of outfitList) {
-    test(`should have all outfits for ${ns}`, async () => {
-      const json = (await import(`../json/${ns}/outfits.json`))
-        .default as OutfitData[];
-      const data = fileList
-        .filter((f) => f.Key[0].startsWith(`${ns}/outfits`))
-        .map((i) => i.Key[0].split("/")[2])
-        .filter((el) => el);
-      for (const file of data) {
-        const outfit = json.find((a) => a.fileName === file);
-        assert.isDefined(outfit, `JSON is missing ${file}`);
-      }
-      for (const object of json) {
-        assert.include(data, object.fileName);
-      }
-      assert.equal(data.length, json.length);
-    });
-
     test(`should have unique names and filenames for ${ns}`, async () => {
       const json = (await import(`../json/${ns}/outfits.json`))
         .default as OutfitData[];
@@ -208,36 +139,7 @@ suite("Asset data", () => {
     });
   }
 
-  test(`should have all picrew`, async () => {
-    const json = (await import(`../json/naomi/picrew.json`)).default;
-    const data = fileList
-      .filter((f) => f.Key[0].startsWith(`naomi/picrew`))
-      .map((i) => i.Key[0].split("/")[2])
-      .filter((el) => el);
-    for (const file of data) {
-      const picrew = json.find((a) => a === file);
-      assert.isDefined(picrew, `JSON is missing ${file}`);
-    }
-  });
-
   for (const ns of portraitsList) {
-    test(`should have all portraits for ${ns}`, async () => {
-      const json = (await import(`../json/${ns}/portraits.json`))
-        .default as PortraitsData[];
-      const data = fileList
-        .filter((f) => f.Key[0].startsWith(`${ns}/art`))
-        .map((i) => i.Key[0].split("/")[2])
-        .filter((el) => el);
-      for (const file of data) {
-        const portrait = json.find((a) => a.fileName === file);
-        assert.isDefined(portrait, `JSON is missing ${file}`);
-      }
-      for (const object of json) {
-        assert.include(data, object.fileName);
-      }
-      assert.equal(data.length, json.length);
-    });
-
     test(`should have unique names and filenames for ${ns}`, async () => {
       const json = (await import(`../json/${ns}/portraits.json`))
         .default as PortraitsData[];
@@ -279,23 +181,6 @@ suite("Asset data", () => {
   }
 
   for (const ns of posesList) {
-    test(`should have all poses for ${ns}`, async () => {
-      const json = (await import(`../json/${ns}/poses.json`))
-        .default as PosesData[];
-      const data = fileList
-        .filter((f) => f.Key[0].startsWith(`${ns}/koikatsu`))
-        .map((i) => i.Key[0].split("/")[2])
-        .filter((el) => el);
-      for (const file of data) {
-        const pose = json.find((a) => a.fileName === file);
-        assert.isDefined(pose, `JSON is missing ${file}`);
-      }
-      for (const object of json) {
-        assert.include(data, object.fileName);
-      }
-      assert.equal(data.length, json.length);
-    });
-
     test(`should have unique names and filenames for ${ns}`, async () => {
       const json = (await import(`../json/${ns}/poses.json`))
         .default as PosesData[];
